Modified demo/demo.py to load a buffer of +/- 10 files surrounding the current file rather than loading all into the DOM at the start.
Also load and draw the first graph before filling the rest of the
buffer.

Demo used to take a very long time to start if run on many (hundreds
or thousands of bamcheck files).  Can now be run on thousands of
files without issue.

Implementation is very simplistic and coupled to the next/prev
navigation model. If using a UI widget for selecting files,
one would probably want to couple the buffering with that.